requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.2.5
honeycomb-beeline==2.9.1
idna==2.8
libhoney==1.9.0
psycopg2==2.8.3
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
statsd==3.3.0
urllib3==1.25.6
wrapt==1.11.2
